contracts/inventory-service/requirements-test.txt
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
jsonschema==4.21.1
pyjwt==2.8.0
faker==24.3.0